fix file size limit preference

+++ app/src/main/kotlin/app/atomofiron/searchboxapp/screens/finder/adapter/holder/EditMaxSizeHolder.kt
@@ -21,7 +21,7 @@ class EditMaxSizeHolder(
     init {
         binding.box.setHint(R.string.pref_max_size)
         binding.field.maxLines = 1
-        binding.field.makeByteSize { output.onEditMaxSize(it.toLong()) }
+        binding.field.makeByteSize { output.onEditMaxSize(it) }
         binding.field.makeHoled(binding.box)
     }
